+++ src/main/java/roomescape/reservation/dto/response/MyReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
