+++ src/main/java/roomescape/reservation/dto/response/PopularThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.response;
 
 import roomescape.reservation.domain.Theme;
