+++ src/test/java/roomescape/reservation/repository/ReservationRepositoryTest.java
@@ -15,7 +15,7 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.domain.Theme;
-import roomescape.reservation.dto.ReservationSearchRequest;
+import roomescape.reservation.dto.request.ReservationSearchRequest;
 
 @DataJpaTest
 public class ReservationRepositoryTest {
